fix some build errors Fingerprint uses String internally again
