+++ keychain/rustfmt.toml
@@ -1,0 +1,3 @@
+hard_tabs = true
+wrap_comments = true
+write_mode = "Overwrite"
